大华项目代码/dahua-master/requirement.txt
Click==7.0
Flask==1.1.1
Flask-RESTful==0.3.7
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==6.2.1
Werkzeug==0.16.0
aniso8601==8.0.0
certifi==2019.11.28
chardet==3.0.4
idna==2.8
itsdangerous==1.1.0
pip==9.0.1
pycrypto==2.6.1
pytz==2019.3
requests==2.22.0
selenium==3.141.0
setuptools==28.8.0
six==1.13.0
urllib3==1.25.7
zxing==0.9.3
xlrd==1.2.0
PyMySQL==0.9.3
PyMuPDF==1.16.10
sqlalchemy==1.3.12
flask_sqlalchemy==2.4.1
